+++ src/index.js
@@ -1,16 +1,16 @@
-import Menu from '/src/menu/menu.js';
-import StationManager from '/src/station/station-manager.js';
-import LineManager from '/src/line/line-manager.js';
-import SectionManager from '/src/section/section-manager.js';
-import MapManager from '/src/map/map-manager.js';
+import MenuController from '/src/menu/controller/menu-controller.js';
+import StationController from '/src/station/controller/station-controller.js';
+import LineController from '/src/line/controller/line-controller.js';
+import SectionController from '/src/section/controller/section-controller.js';
+import MapController from '/src/map/controller/map-controller.js';
 
 export default class SubwayStationApp {
 	constructor() {
-		new Menu();
-		new StationManager();
-		new LineManager();
-		new SectionManager();
-		new MapManager();
+		new MenuController();
+		new StationController();
+		new LineController();
+		new SectionController();
+		new MapController();
 	}
 }
