+++ src/map/controller/map-controller.js
@@ -1,0 +1,7 @@
+import MapOutput from '../view/output.js';
+
+export default class MapController {
+	constructor() {
+		this.mapOutput = new MapOutput();
+	}
+}
