--- src/map/map-manager.js
@@ -1,7 +1,0 @@
-import MapOutput from './view/output.js';
-
-export default class MapManager {
-	constructor() {
-		this.mapOutput = new MapOutput();
-	}
-}
